+++ ahyeon/example/public/index.html
@@ -1,11 +1,15 @@
 <!DOCTYPE html>
 <html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <title>ahyeon</title>
-    <script type="module" src="/main.js"></script>
-  </head>
-  <body>
-    <div id="webgl-container"></div>
-  </body>
+
+<head>
+  <meta charset="UTF-8">
+  <title>#03. Geometry</title>
+  <link rel="stylesheet" href="../src/style.css">
+  <script type="module" src="/main.js"></script>
+</head>
+
+<body>
+  <div id="webgl-container"></div>
+</body>
+
 </html>
