--- ahyeon/example/src/main.ts
@@ -1,120 +1,0 @@
-import * as THREE from "three";
-// import { Vector3 } from "three";
-import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
-import { FontLoader } from "three/examples/jsm/loaders/FontLoader.js"
-import { TextGeometry } from "three/examples/jsm/geometries/TextGeometry.js"
-
-
-class App {
-  private _divContainer: HTMLElement;
-  private _renderer: THREE.WebGLRenderer;
-  private _scene: THREE.Scene;
-  private _camera!: THREE.PerspectiveCamera;
-  // private _cube!: THREE.Mesh | THREE.Group;
-
-  constructor() {
-    const divContainer = document.querySelector("#webgl-container") as HTMLElement;
-    this._divContainer = divContainer;
-
-    const renderer = new THREE.WebGLRenderer({ antialias: true });
-    this._renderer = renderer;
-    renderer.setPixelRatio(window.devicePixelRatio);
-
-    divContainer.appendChild(renderer.domElement);
-
-    const scene = new THREE.Scene();
-    this._scene = scene;
-
-    this._setupCamera();
-    this._setupLight();
-    this._setupModel();
-    this._setupControls();
-
-    window.onresize = this.resize.bind(this);
-    this.resize();
-
-    requestAnimationFrame(this.render.bind(this));
-  }
-
-  private _setupCamera() {
-    const width = this._divContainer.clientWidth;
-    const height = this._divContainer.clientHeight;
-    const camera = new THREE.PerspectiveCamera(75, width / height, 0.1, 100);
-    camera.position.z = 10;
-    this._camera = camera;
-  }
-
-  private _setupLight() {
-    const color = 0xffffff;
-    const intensity = 1;
-    const light = new THREE.DirectionalLight(color, intensity);
-    light.position.set(-1, 2, 4);
-    this._scene.add(light);
-  }
-
-  private _setupModel() {
-    const fontLoader = new FontLoader();
-    async function loadFonts(that: any) {
-      const url = "../examples/fonts/Pretendard-Regular";
-      const font = await new Promise<any>((resolve, reject) => {
-        fontLoader.load(url, resolve, undefined, reject);
-      })
-
-      const geometry = new TextGeometry("test", {
-        font: font,
-        size: 5,
-        height: 1,
-        curveSegments: 4,
-        bevelEnabled: true,
-        bevelThickness: 0.5,
-        bevelSize: .5,
-        bevelOffset: 0,
-        bevelSegments: 2
-      })
-
-
-      const material = new THREE.MeshPhongMaterial({ color: 0x515151 });
-      const cube = new THREE.Mesh(geometry, material);
-
-      const lineMaterial = new THREE.LineBasicMaterial({ color: 0xffffff });
-      const line = new THREE.LineSegments(new THREE.WireframeGeometry(geometry), lineMaterial);
-
-      const group = new THREE.Group()
-      group.add(cube);
-      group.add(line);
-
-      that._scene.add(group);
-      that._cube = group;
-    }
-
-    loadFonts(this);
-  }
-
-  private _setupControls() {
-    new OrbitControls(this._camera, this._divContainer); // camera,webgl_container
-  }
-
-  resize() {
-    const width = this._divContainer.clientWidth;
-    const height = this._divContainer.clientHeight;
-
-    this._camera.aspect = width / height;
-    this._camera.updateProjectionMatrix();
-
-    this._renderer.setSize(width, height);
-  }
-
-  render() {
-    this._renderer.render(this._scene, this._camera);
-    // this.update(time);
-    requestAnimationFrame(this.render.bind(this));
-  }
-
-  // update(time: number) {
-  //   time *= 0.0001; // secondunit
-  // }
-}
-
-window.onload = function () {
-  new App();
-};
